Fix printf type casts in Pixel DMX code
Add explicit `static_cast<unsigned>` conversions for `%u` formatting in Pixel DMX configuration, multi-port control, JSON status output, and parameter dump/setup paths to avoid signedness/type warnings and improve correctness. This also includes small readability cleanups such as renaming a temporary brightness variable and simplifying start universe assignment.

+++ lib-pixeldmx/include/pixeldmxmulti.h
@@ -96,7 +96,7 @@ class PixelDmxMulti final : public PixelDmxConfiguration {
 
     void Start(uint32_t port_index) {
         const auto kIndex = (port_index <= 31) ? 0 : 1;
-        DEBUG_PRINTF("%u [%u]", port_index, kIndex);
+        DEBUG_PRINTF("%u [%u]", static_cast<unsigned>(port_index), static_cast<unsigned>(kIndex));
 
         if (kIndex == 0) {
             started_[0] |= (1U << port_index);
@@ -113,7 +113,7 @@ class PixelDmxMulti final : public PixelDmxConfiguration {
 
     void Stop(uint32_t port_index) {
         const auto kIndex = (port_index <= 31) ? 0 : 1;
-        DEBUG_PRINTF("%u [%u]", port_index, kIndex);
+        DEBUG_PRINTF("%u [%u]", static_cast<unsigned>(port_index), static_cast<unsigned>(kIndex));
 
         if (kIndex == 0) {
             if (started_[0] & (1U << port_index)) {
